requirements.txt
GitPython==3.1.43
gymnasium==1.0.0
numpy==2.1.3
tensorboardX==2.6.2.2
torch==2.5.0
tqdm==4.66.5
